Flytt barn-typer inn i egen barn-fil
