+++ src/frontend/components/SøknadsSteg/OmBarnaDine/HvilkeBarnCheckboxGruppe.tsx
@@ -9,8 +9,8 @@ import { ESvar } from '@navikt/familie-form-elements';
 import { Felt } from '@navikt/familie-skjema';
 
 import { useApp } from '../../../context/AppContext';
+import { barnDataKeySpørsmål } from '../../../typer/barn';
 import { BarnetsId } from '../../../typer/common';
-import { barnDataKeySpørsmål } from '../../../typer/person';
 import { barnetsNavnValue } from '../../../utils/barn';
 import SpråkTekst from '../../Felleskomponenter/SpråkTekst/SpråkTekst';
