+++ src/frontend/components/SøknadsSteg/OmBarnaDine/OmBarnaDine.tsx
@@ -7,7 +7,7 @@ import { ESvar } from '@navikt/familie-form-elements';
 
 import { useApp } from '../../../context/AppContext';
 import { useEøs } from '../../../context/EøsContext';
-import { barnDataKeySpørsmål } from '../../../typer/person';
+import { barnDataKeySpørsmål } from '../../../typer/barn';
 import AlertStripe from '../../Felleskomponenter/AlertStripe/AlertStripe';
 import JaNeiSpm from '../../Felleskomponenter/JaNeiSpm/JaNeiSpm';
 import KomponentGruppe from '../../Felleskomponenter/KomponentGruppe/KomponentGruppe';
