+++ src/frontend/components/SøknadsSteg/OmBarnaDine/useBarnCheckboxFelt.tsx
@@ -4,8 +4,8 @@ import { ESvar } from '@navikt/familie-form-elements';
 import { Avhengigheter, feil, Felt, FeltState, ok, useFelt } from '@navikt/familie-skjema';
 
 import { useApp } from '../../../context/AppContext';
+import { barnDataKeySpørsmål } from '../../../typer/barn';
 import { BarnetsId } from '../../../typer/common';
-import { barnDataKeySpørsmål } from '../../../typer/person';
 import SpråkTekst from '../../Felleskomponenter/SpråkTekst/SpråkTekst';
 
 const useBarnCheckboxFelt = (
