+++ src/frontend/components/SøknadsSteg/VelgBarn/useVelgBarn.tsx
@@ -5,10 +5,10 @@ import { feil, ISkjema, ok, useFelt, useSkjema } from '@navikt/familie-skjema';
 import { useApp } from '../../../context/AppContext';
 import { useEøs } from '../../../context/EøsContext';
 import { useSteg } from '../../../context/StegContext';
+import { IBarnMedISøknad } from '../../../typer/barn';
 import { BarnetsId } from '../../../typer/common';
 import { IBarn } from '../../../typer/person';
 import { IVelgBarnFeltTyper } from '../../../typer/skjema';
-import { IBarnMedISøknad } from '../../../typer/søknad';
 import { setUserProperty, UserProperty } from '../../../utils/amplitude';
 import { genererInitialBarnMedISøknad } from '../../../utils/barn';
 import SpråkTekst from '../../Felleskomponenter/SpråkTekst/SpråkTekst';
