+++ src/frontend/context/StegContext.tsx
@@ -4,8 +4,8 @@ import createUseContext from 'constate';
 import { StegindikatorStegProps } from 'nav-frontend-stegindikator/lib/stegindikator-steg';
 import { matchPath, useLocation } from 'react-router-dom';
 
+import { IBarnMedISøknad } from '../typer/barn';
 import { ISteg, RouteEnum } from '../typer/routes';
-import { IBarnMedISøknad } from '../typer/søknad';
 import { useApp } from './AppContext';
 import { useEøs } from './EøsContext';
 import { useRoutes } from './RoutesContext';
